+++ DesktopApplicationTemplate.UI/Views/Ftp/FTPServiceView.xaml.cs
@@ -21,7 +21,7 @@ namespace DesktopApplicationTemplate.UI.Views.Ftp
 
         public void SetServiceContext(ServiceListModel service)
         {
-            LogView.DataContext = new ServiceLogViewModel(service.ServiceType, service.Logs);
+            LogView.DataContext = new ServiceLogViewModel(service.Type, service.Logs);
         }
     }
 }